_collections + _opcode: extract into module/_collections/ + module/_opcode/
importing.rs shrinks by 439 lines.
