+++ pyre/pyre-interpreter/src/module/_collections/moduledef.rs
@@ -1,0 +1,7 @@
+//! _collections module definition — PyPy: pypy/module/_collections/moduledef.py
+
+use crate::DictStorage;
+
+pub fn init(ns: &mut DictStorage) {
+    super::interp_collections::register_module(ns);
+}
