+++ pyre/pyre-interpreter/src/module/_opcode/mod.rs
@@ -1,0 +1,7 @@
+//! _opcode module — PyPy: pypy/module/_opcode/
+//!
+//! Stub providing stack_effect + has_arg/has_const/has_name/has_jump and
+//! related classifiers — opcode.py requirements.  Returns neutral values.
+
+pub mod interp_opcode;
+pub mod moduledef;
